requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlmodel>=0.0.22
spotipy>=2.24.0
rapidfuzz>=3.10.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
jinja2>=3.1.0
python-multipart>=0.0.18
python-dotenv>=1.0.0
